requirements.txt
mss==9.0.1
pillow==10.4.0
pynput==1.7.7
pystray==0.19.5
six==1.16.0
